# swagger_Timebank_v0.4.yaml
openapi: 3.0.1
info:
  title: Timebank
  description: 'This is API description for Timebank project.'
  version: 0.4.0
servers:
- url: http://localhost:5000/api/v1
...
paths:
  /users:
    get:
      security: 
      - bearerAuth: []
      tags:
      - users
      summary: Get all users
      description: Get all users from the database
      parameters:
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/sortFieldParam'
      responses:
        200:
          description: OK
          content: {}
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /user/{id}:
    parameters:
      - name: id
        in: path
        description: ID of user to use
        required: true
        schema:
          type: integer
          format: int64
    get:
        security:
        - bearerAuth: []
        tags:
        - users
        summary: Find user by ID
        description: Returns a single user
        operationId: getUserById
        responses:
          200:
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/UsersWithServices'
          400:
            $ref: '#/components/responses/BadParam'
          401:
            $ref: '#/components/responses/NotAuthorized'
          404:
            $ref: '#/components/responses/NotFound'
          5XX:
            $ref: '#/components/responses/ServerError'
    put:
      security:
      - bearerAuth: []
      tags:
      - users
      summary: Updates the user
      operationId: updateUser
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserWithPasswordUpdate'
          application/json:
            schema:
              $ref: '#/components/schemas/UserWithPasswordUpdate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
    delete:
      security:
      - bearerAuth: []
      tags:
      - users
      summary: Delete user
      operationId: deleteUser
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'  
  /user-create:
    post:
      security:
      - bearerAuth: []
      tags:
        - users
      summary: Create a user in the library
      operationId: createUser
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserWithPassword'
          application/json:
            schema:
              $ref: '#/components/schemas/UserWithPassword'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /user/{id}/set-password:
    parameters:
      - name: id
        in: path
        description: ID of user to set
        required: true
        schema:
          type: integer
          format: int64
    put:
      security:
      - bearerAuth: []
      tags:
      - users
      summary: Set new password to user
      operationId: setUserPassword
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserPassword'
          application/json:
            schema:
              $ref: '#/components/schemas/UserPassword'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /user/login:
    post:
      security: []
      tags:
        - users
      summary: Login as user
      operationId: userLogin
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLogin'
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /user/logout:
    post:
      security:
      - bearerAuth: []
      tags:
        - users
      summary: Logout user
      operationId: userLogout
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /user/profile:
    get:
      security:
      - bearerAuth: []
      tags:
        - users
      summary: Get logged user details
      operationId: userProfile
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    Users:
      required:
        - user_name
        - phone
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_name:
          type: string
          example: Janko Hrasko
        phone:
          type: string
          example: +421000111222
        time_account:
          type: integer
          format: int64
          example: 1
    UsersUpdate:
      type: object
      properties:
        user_name:
          type: string
          example: Janko Hrasko
        phone:
          type: string
          example: +421000111222
        time_account:
          type: integer
          format: int64
          example: 1
    UsersWithServices:
        allOf:
          - $ref: '#/components/schemas/Users'
          - type: object
            properties:
              services:
                $ref: '#/components/schemas/Service'  
    UserWithPassword:
      allOf:
        - $ref: '#/components/schemas/Users'
        - type: object
          properties:
            password:
              type: string
              format: password
              example: 9G(6`C2i~S]
        - type: object
          properties:
            password_val:
              type: string
              format: password
              example: 9G(6`C2i~S]
    UserWithPasswordUpdate:
      allOf:
        - $ref: '#/components/schemas/UsersUpdate'
        - type: object
          properties:
            password:
              type: string
              format: password
              example: janko
    UserPassword:
      required:
        - password
        - password_val
      type: object
      properties:
        password:
          type: string
          format: password
          example: 9G(6`C2i~S]
        password_val:
          type: string
          format: password
          example: 9G(6`C2i~S]
    UserLogin:
      required:
        - phone
        - password
      type: object
      properties:
        phone:
          type: string
          example: +421000111222
        password:
          type: string
          format: password
          example: kastan
    Service:
      required:
        - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Pokosim travnik
        estimate:
          type: integer
          format: int64
          example: 2
...
  parameters:
    sortOrderParam:
      in: query
      name: sort
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
      description: Used to sort return values in ascending or descending order
    sortFieldParam:
      in: query
      name: field
      required: false
      schema:
        type: string
      description: Used to select sorting field from the model
...
  responses:
    NotFound:
      description: No entries found.
    NotAuthorized:
      description: Authorization information is missing or invalid.
    BadParam:
      description: Bad paramters.
    ServerError:
      description: Unexpected response error. Server fault.
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: bearer
      bearerFormat: JWT